+++ healthcare/gk/views.py
@@ -705,11 +705,11 @@ class AddActionForm_GoalPage(forms.Form):
 	
 
 class AddQuantStatusForm(forms.Form):
-	data_Value = forms.IntegerField()
+	data_Value = forms.FloatField()
 	notes = forms.CharField()
 
 class UpdateQuantStatusForm(forms.Form):
-	data_Value = forms.IntegerField()
+	data_Value = forms.FloatField()
 	notes = forms.CharField()
 	pub_time = forms.DateTimeField(input_formats=["%m/%d %Y %H:%M"])
